--- docs/reference/django_celery_monitor.admin_utils.rst
@@ -1,11 +1,0 @@
-=======================================
- ``django_celery_monitor.admin_utils``
-=======================================
-
-.. contents::
-    :local:
-.. currentmodule:: django_celery_monitor.admin_utils
-
-.. automodule:: django_celery_monitor.admin_utils
-    :members:
-    :undoc-members:
